split script for plotting megadudes qc into calculating and plotting script, add tests
