+++ test/rules/qc_megadudes/calculate_megadudes_qc_metrics_simulation/data/config/config.yaml
@@ -1,0 +1,3 @@
+alignment_methods: [
+  "diamond"
+]
