+++ test/rules/qc_megadudes/calculate_megadudes_qc_metrics_simulation/data/results/diamond/simulated_peptides_0_with_1_percent_noise-lineage.tsv
@@ -1,0 +1,10 @@
+superkingdom	phylum	class	order	family	genus	species	subspecies
+2.0	2818505.0		3031712.0	49.0	39643.0	56	
+2.0	1224.0	28211.0	2800060.0	69657.0	85.0	87	
+2.0	203691.0	203692.0	136.0	137.0	146.0	150	
+2.0	203691.0	203692.0	136.0	2845253.0	157.0	158	
+2.0	203691.0	203692.0	136.0	2845253.0	157.0	160	
+2.0	203691.0	203692.0	136.0	2845253.0	157.0	162	
+2.0	203691.0	203692.0	136.0	2845253.0	157.0	166	
+2.0	29547.0	3031852.0	213849.0	72294.0	194.0	195	
+2.0	29547.0	3031852.0	213849.0	72294.0	194.0	197
