+++ test/rules/qc_megadudes/calculate_megadudes_qc_metrics_simulation/data/results/simulation/sample_taxons_lineage_0.tsv
@@ -1,0 +1,3 @@
+Kleiner et al. Name	superkingdom	phylum	class	order	family	genus	species	subspecies	Protein amount in equal Protein [microg]
+Agrobacterium tumefaciens NTL4	2	1224	28211	356	82115	357	358		642.7
+Alteromonas macleodii ATCC 27126	2	1224	1236	135622	72275	226	28108	1004788	642.7
