+++ test/rules/qc_megadudes/calculate_megadudes_qc_metrics_simulation/data/results/unipept/simulated_peptides_0_with_1_percent_noise.csv
@@ -1,0 +1,10 @@
+peptide,taxon_id,taxon_name,taxon_rank,domain_id,superkingdom_name,kingdom_id,kingdom_name,subkingdom_id,subkingdom_name,superphylum_id,superphylum_name,phylum_id,phylum_name,subphylum_id,subphylum_name,superclass_id,superclass_name,class_id,class_name,subclass_id,subclass_name,superorder_id,superorder_name,order_id,order_name,suborder_id,suborder_name,infraorder_id,infraorder_name,superfamily_id,superfamily_name,family_id,family_name,subfamily_id,subfamily_name,tribe_id,tribe_name,subtribe_id,subtribe_name,genus_id,genus_name,subgenus_id,subgenus_name,species_group_id,species_group_name,species_subgroup_id,species_subgroup_name,species_id,species_name,subspecies_id,subspecies_name,strain_id,strain_name,varietas_id,varietas_name,forma_id,forma_name
+FNIMYGTGISK,1129771,Leptotrichiaceae,family,2,Bacteria,,,,,,,32066,Fusobacteriota,,,,,203490,Fusobacteriia,,,,,203491,Fusobacteriales,,,,,,,1129771,Leptotrichiaceae,,,,,,,,,,,,,,,,,,,,,,,,
+AEIEGEMGDQQMGLQAR,2,Bacteria,superkingdom,2,Bacteria,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+AGGVVAFIDAEHALDPVYAK,40542,Leptotrichia buccalis,species,2,Bacteria,,,,,,,32066,Fusobacteriota,,,,,203490,Fusobacteriia,,,,,203491,Fusobacteriales,,,,,,,1129771,Leptotrichiaceae,,,,,,,32067,Leptotrichia,,,,,,,40542,Leptotrichia buccalis,,,,,,,,
+QGDDVVGSEVVVK,40542,Leptotrichia buccalis,species,2,Bacteria,,,,,,,32066,Fusobacteriota,,,,,203490,Fusobacteriia,,,,,203491,Fusobacteriales,,,,,,,1129771,Leptotrichiaceae,,,,,,,32067,Leptotrichia,,,,,,,40542,Leptotrichia buccalis,,,,,,,,
+SGAIDVIVVDSVAALVPK,2,Bacteria,superkingdom,2,Bacteria,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+IGGFSFVPGVQTTTSGGR,1129771,Leptotrichiaceae,family,2,Bacteria,,,,,,,32066,Fusobacteriota,,,,,203490,Fusobacteriia,,,,,203491,Fusobacteriales,,,,,,,1129771,Leptotrichiaceae,,,,,,,,,,,,,,,,,,,,,,,,
+AISTGSLNLDIALGIGGVPR,1129771,Leptotrichiaceae,family,2,Bacteria,,,,,,,32066,Fusobacteriota,,,,,203490,Fusobacteriia,,,,,203491,Fusobacteriales,,,,,,,1129771,Leptotrichiaceae,,,,,,,,,,,,,,,,,,,,,,,,
+ALGVNIDELLISQPDTGEQALEISDMLVR,32067,Leptotrichia,genus,2,Bacteria,,,,,,,32066,Fusobacteriota,,,,,203490,Fusobacteriia,,,,,203491,Fusobacteriales,,,,,,,1129771,Leptotrichiaceae,,,,,,,32067,Leptotrichia,,,,,,,,,,,,,,,,
+SDTVMIFINQIR,1129771,Leptotrichiaceae,family,2,Bacteria,,,,,,,32066,Fusobacteriota,,,,,203490,Fusobacteriia,,,,,203491,Fusobacteriales,,,,,,,1129771,Leptotrichiaceae,,,,,,,,,,,,,,,,,,,,,,,,
