+++ test/rules/qc_megadudes/calculate_megadudes_qc_metrics_simulation/expected/results/qc/qc-simulated_peptides_0_with_1_percent_noise.tsv
@@ -1,0 +1,22 @@
+	method	taxon_level	Sensitivity	Precision	F1-score	FDR	taxon_abbreviation
+0	DIAMOND	superkingdom	100.0	100.0	100.0	0.0	SK
+1	Unipept	superkingdom	100.0	100.0	100.0	0.0	SK
+2	DUDes	superkingdom	100.0	100.0	100.0	0.0	SK
+3	DIAMOND	phylum	100.0	25.0	40.0	75.0	P
+4	Unipept	phylum	0.0	0.0		100.0	P
+5	DUDes	phylum	0.0	0.0		100.0	P
+6	DIAMOND	class	50.0	33.333333333333336	40.0	66.66666666666667	C
+7	Unipept	class	0.0	0.0		100.0	C
+8	DUDes	class	0.0				C
+9	DIAMOND	order	0.0	0.0		100.0	O
+10	Unipept	order	0.0	0.0		100.0	O
+11	DUDes	order	0.0				O
+12	DIAMOND	family	0.0	0.0		100.0	F
+13	Unipept	family	0.0	0.0		100.0	F
+14	DUDes	family	0.0				F
+15	DIAMOND	genus	0.0	0.0		100.0	G
+16	Unipept	genus	0.0	0.0		100.0	G
+17	DUDes	genus	0.0				G
+18	DIAMOND	species	0.0	0.0		100.0	S
+19	Unipept	species	0.0	0.0		100.0	S
+20	DUDes	species	0.0				S
