+++ test/rules/qc_megadudes/plot_megadudes_qc/data/config/config.yaml
@@ -1,0 +1,3 @@
+alignment_methods: [
+  "diamond"
+]
