+++ test/rules/qc_megadudes/plot_megadudes_qc/test_plot_megadudes_qc.py
@@ -1,9 +1,9 @@
 from test.common import snakemake_run, check_output
 
 
-def test_plot_alignment_qc(tmpdir, workflow_path, prepared_workdir):
-    targets = ["plots/alignment/qc-foo.svg"]
-    snakefile = workflow_path / "workflow/rules/qc_alignment.smk"
+def test_plot_megadudes_qc(tmpdir, workflow_path, prepared_workdir):
+    targets = ["plots/megadudes/qc-foo.svg"]
+    snakefile = workflow_path / "workflow/rules/qc_megadudes.smk"
     configfile = prepared_workdir.workdir / "config" / "config.yaml"
     conda_prefix = workflow_path / ".snakemake" / "conda"
 
@@ -14,10 +14,10 @@ def test_plot_alignment_qc(tmpdir, workflow_path, prepared_workdir):
         prepared_workdir.workdir,
         additional_arguments=[
             "--use-conda",
-            "--configfile",
-            configfile.as_posix(),
             "--conda-prefix",
             conda_prefix.as_posix(),
+            "--configfile",
+            configfile.as_posix(),
         ]
     )
